requirements.txt
aiofiles==0.7.0
aiogrpc==1.8
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
async-lru==1.0.3
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
click==7.1.2
cryptography==3.4.7
fastapi==0.63.0
frozenlist==1.2.0
grpcio==1.38.0
grpcio-tools==1.38.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
humanfriendly==9.1
loguru==0.5.3
matplotlib-inline==0.1.2
multidict==5.2.0
nmcli==0.5.0
paramiko==2.7.2
passlib==1.7.4
protobuf==3.17.1
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.2
pyreadline==2.1
PySocks==1.7.1
python-multipart==0.0.5
pyudev==0.22.0
rfc3986==1.5.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.13.4
websockets==9.1
xmltodict==0.12.0
